--- src/main/java/com/boaz/backend/domain/auth/dto/response/UserTokenRefreshResponse.java
@@ -1,7 +1,0 @@
-package com.boaz.backend.domain.auth.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record UserTokenRefreshResponse(
-        @JsonProperty("access_token") String accessToken
-) {}
