+++ src/main/java/com/boaz/backend/global/oauth/CookieOAuth2AuthorizationRequestRepository.java
@@ -1,6 +1,7 @@
 package com.boaz.backend.global.oauth;
 
 import com.boaz.backend.global.util.CookieProvider;
+import com.boaz.backend.global.util.CookieUtils;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
